fix(quality): move the shared identity mutation helper into canonical_pipeline/identity_fields.py so speaker.py stays under its shrink-only size baseline
