+++ backend/utils/canonical_pipeline/speaker.py
@@ -1,4 +1,5 @@
 from .constants import *
+from .identity_fields import apply_recording_speaker_identity_fields
 
 
 @dataclass(frozen=True)
@@ -1076,56 +1077,6 @@ def record_recording_speaker_corrections(
         )
 
     return correction_events
-
-
-def apply_recording_speaker_identity_fields(
-    session,
-    recording_speaker: RecordingSpeaker,
-    *,
-    new_speaker_name: str,
-    target_global_speaker: GlobalSpeaker | None = None,
-    merge_global_embedding_alpha: float | None = None,
-    identity_confidence: float | None = None,
-    identity_locked: bool | None = None,
-    sync_aliases: bool = True,
-) -> None:
-    """Mutate one recording speaker's identity fields in place.
-
-    Shared core between the manual rename path
-    (``update_recording_speaker_identity``) and the worker's speaker-suggestion
-    auto-apply. Voiceprint-locked global speakers never receive embedding
-    merges. ``identity_confidence`` and ``identity_locked`` are written only
-    when a value is provided, so callers control whether the identity is
-    asserted as human-confirmed.
-    """
-    if target_global_speaker is not None:
-        recording_speaker.global_speaker_id = target_global_speaker.id
-        recording_speaker.global_speaker = target_global_speaker
-        recording_speaker.local_name = None
-        if merge_global_embedding_alpha is not None and recording_speaker.embedding:
-            if target_global_speaker.embedding:
-                if not target_global_speaker.is_voiceprint_locked:
-                    target_global_speaker.embedding = merge_embeddings(
-                        target_global_speaker.embedding,
-                        recording_speaker.embedding,
-                        alpha=merge_global_embedding_alpha,
-                    )
-            else:
-                target_global_speaker.embedding = list(recording_speaker.embedding)
-            session.add(target_global_speaker)
-    else:
-        recording_speaker.global_speaker_id = None
-        recording_speaker.global_speaker = None
-        recording_speaker.local_name = new_speaker_name
-
-    recording_speaker.name = None
-    if identity_confidence is not None:
-        recording_speaker.identity_confidence = identity_confidence
-    if identity_locked is not None:
-        recording_speaker.identity_locked = identity_locked
-    if sync_aliases:
-        ensure_recording_speaker_aliases_for_speaker(session, recording_speaker)
-    session.add(recording_speaker)
 
 
 def update_recording_speaker_identity(
